+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Implement `MonitorId::get_dimensions` for Android.
+
 # Version 0.10.1 (2018-02-05)
 
 - Added method `os::macos::WindowBuilderExt::with_movable_by_window_background(bool)` that allows to move a window without a titlebar - `with_decorations(false)`